+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 4. Created `StringAlignUtils`
 5. Added option to not render gray background in ScreenUIRenderer
 6. Added `sizeBehaviour` field to `Container`
+7. Fixed `DynamicUIComponent` crashes
 
 # v1.0.5 - Minor Update
 ## Built-in Components